+++ custom_components/supernotify/const.py
@@ -92,6 +92,7 @@ OCCUPANCY_ONLY_IN = "only_in"
 OCCUPANCY_ONLY_OUT = "only_out"
 
 ATTR_ENABLED = "enabled"
+ATTR_TRANSPORT_ENABLED = "transport_enabled"
 ATTR_PRIORITY = "priority"
 ATTR_ACTION = "action"
 ATTR_SCENARIOS_REQUIRE = "require_scenarios"